refactored extenders from scheduler/core to scheduler
